+++ src/test/java/io/github/astrapi69/lang/ClassExtensionsTest.java
@@ -1,17 +1,17 @@
 /**
  * The MIT License
- *
+ * <p>
  * Copyright (C) 2015 Asterios Raptis
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
  * associated documentation files (the "Software"), to deal in the Software without restriction,
  * including without limitation the rights to use, copy, modify, merge, publish, distribute,
  * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in all copies or
  * substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
  * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
  * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
@@ -37,7 +37,11 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Properties;
+import java.util.Set;
 
+import com.google.common.reflect.ClassPath;
+import io.github.astrapi69.collection.set.SetFactory;
+import io.github.astrapi69.throwable.RuntimeExceptionDecorator;
 import net.sf.cglib.proxy.Enhancer;
 import net.sf.cglib.proxy.MethodInterceptor;
 import net.sf.cglib.proxy.MethodProxy;
@@ -105,6 +109,51 @@ public class ClassExtensionsTest
 		expected = true;
 		assertEquals(expected, actual);
 
+	}
+
+	/**
+	 * Test method for {@link ClassExtensions#isInstantiable(Class)} from all classes in the
+	 * classloader
+	 */
+	@Test
+	public void testIsInstantiableWithAllClasses()
+	{
+		boolean expected;
+		boolean actual;
+		final ClassLoader classLoader = ClassExtensions.getClassLoader(ClassExtensions.class);
+		ClassPath classPath = RuntimeExceptionDecorator.decorate(() -> ClassPath.from(classLoader));
+		Set<ClassPath.ClassInfo> classes = classPath.getAllClasses();
+		Set<Class> notInstantiableClasses = SetFactory.newHashSet();
+		Set<Class> instantiableClasses = SetFactory.newHashSet();
+		Set<String> notLoadableClasses = SetFactory.newHashSet();
+		assertEquals(classes.size(), 8046);
+
+		classes.stream()
+			.filter(classInfo -> classInfo.getPackageName().startsWith("io.github.astrapi69"))
+			.forEach(classInfo -> {
+				Class<?> load;
+				try
+				{
+					load = classInfo.load();
+					final boolean instantiable = ClassExtensions.isInstantiable(load);
+					if (instantiable)
+					{
+						instantiableClasses.add(load);
+					}
+					else
+					{
+						notInstantiableClasses.add(load);
+					}
+
+				}
+				catch (Throwable t)
+				{
+					notLoadableClasses.add(classInfo.getName());
+				}
+			});
+		assertEquals(notInstantiableClasses.size(), 44);
+		assertEquals(instantiableClasses.size(), 200);
+		assertEquals(notLoadableClasses.size(), 0);
 	}
 
 	/**
@@ -776,7 +825,6 @@ public class ClassExtensionsTest
 			ClassLoader.getSystemClassLoader());
 		expected = true;
 		assertEquals(expected, actual);
-
 
 		actual = ClassExtensions.isDerivate(null, null);
 		assertEquals(expected, actual);